Use a custom "AsyncTextView" for the entity-ui
As noted in the comments, the default TextView can take >100ms to
render when the content is long, such as for the Combat SRD entry.
With our UI, that becomes a very noticeable jank when searching, so
the AsyncTextView will render just enough to fill the current page,
then perform the rest of the rendering asynchronously. This makes
loading even the Combat entry feel instantaneous!