Add some support for bulleted lists in the SRD
Unfortunately the JSON we're using is not super consistent. A sequence
of paragraphs looks the same as a list. Since paragraphs are more
common, we assume that. But, this is giving me more and more reason to
want to rework to base off the open5e rst documents, or something.