Support searching the entire SRD (#12)
* Add initial experiments with making the full SRD searchable

* Don't include empty container rule; fix tests

* Use a custom "AsyncTextView" for the entity-ui

As noted in the comments, the default TextView can take >100ms to
render when the content is long, such as for the Combat SRD entry.
With our UI, that becomes a very noticeable jank when searching, so
the AsyncTextView will render just enough to fill the current page,
then perform the rest of the rendering asynchronously. This makes
loading even the Combat entry feel instantaneous!

* Add some support for bulleted lists in the SRD

Unfortunately the JSON we're using is not super consistent. A sequence
of paragraphs looks the same as a list. Since paragraphs are more
common, we assume that. But, this is giving me more and more reason to
want to rework to base off the open5e rst documents, or something.

* Clean out old data sources, etc.

* Remove hack trying to avoid re-rendering Entity

It's no longer needed for performance with AsyncTextView